+++ InstantGrowTrees/Framework/ModConfig.cs
@@ -1,4 +1,4 @@
-namespace InstantGrowTrees
+namespace InstantGrowTrees.Framework
 {
     /// <summary>The mod configuration model.</summary>
     internal class ModConfig
@@ -7,7 +7,10 @@ namespace InstantGrowTrees
         ** Accessors
         *********/
         /// <summary>Whether fruit trees grow instantly overnight.</summary>
-        public bool FruitTreesInstantGrow { get; set; }
+        public bool FruitTreesInstantGrow { get; set; } = true;
+
+        /// <summary>Whether fruit trees instantly age to iridium quality.</summary>
+        public bool FruitTreesInstantAge { get; set; } = false;
 
         /// <summary>Whether non-fruit trees grow instantly overnight.</summary>
         public bool RegularTreesInstantGrow { get; set; } = true;