+++ SkullCaveSaver/Properties/AssemblyInfo.cs
@@ -1,5 +1,0 @@
-﻿using System.Reflection;
-
-[assembly: AssemblyTitle("SkullCaveSaver")]
-[assembly: AssemblyVersion("1.2.2")]
-[assembly: AssemblyFileVersion("1.2.2")]